Use FPFH by default, factor heat mappers into multiple files.
