+++ rapid_perception/package.xml
@@ -40,7 +40,6 @@
   <!-- Use test_depend for packages you need only for testing: -->
   <!--   <test_depend>gtest</test_depend> -->
   <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>cv_bridge</build_depend>
   <build_depend>geometry_msgs</build_depend>
   <build_depend>image_geometry</build_depend>
   <build_depend>interactive_markers</build_depend>
@@ -55,9 +54,7 @@
   <build_depend>roscpp</build_depend>
   <build_depend>rostest</build_depend>
   <build_depend>sensor_msgs</build_depend>
-  <build_depend>std_msgs</build_depend>
   <build_depend>visualization_msgs</build_depend>
-  <run_depend>cv_bridge</run_depend>
   <run_depend>geometry_msgs</run_depend>
   <run_depend>image_geometry</run_depend>
   <run_depend>interactive_markers</run_depend>
@@ -72,7 +69,6 @@
   <run_depend>roscpp</run_depend>
   <run_depend>rostest</run_depend>
   <run_depend>sensor_msgs</run_depend>
-  <run_depend>std_msgs</run_depend>
   <run_depend>visualization_msgs</run_depend>
   <test_depend>google-mock</test_depend>
   <test_depend>gtest</test_depend>
